+++ src/utils/index.ts
@@ -1,1 +1,2 @@
+export { ValidationError } from 'yup';
 export { WorkflowError } from './errors.js';