+++ tests/e2e/sendEmail.test.ts
@@ -102,6 +102,7 @@ describe('web3mail.sendEmail()', () => {
     },
     2 * MAX_EXPECTED_BLOCKTIME + MAX_EXPECTED_WEB2_SERVICES_TIME
   );
+
   it(
     'should successfully send email with granted access to whitelist address',
     async () => {
@@ -132,6 +133,7 @@ describe('web3mail.sendEmail()', () => {
     },
     2 * MAX_EXPECTED_BLOCKTIME + MAX_EXPECTED_WEB2_SERVICES_TIME
   );
+
   it(
     'should successfully send email with content type html',
     async () => {
@@ -149,6 +151,7 @@ describe('web3mail.sendEmail()', () => {
     },
     2 * MAX_EXPECTED_BLOCKTIME + MAX_EXPECTED_WEB2_SERVICES_TIME
   );
+
   it(
     'should fail if the protected data is not valid',
     async () => {
@@ -162,12 +165,15 @@ describe('web3mail.sendEmail()', () => {
       await expect(web3mail.sendEmail(params)).rejects.toThrow(
         new WorkflowError({
           message: 'Failed to sendEmail',
-          errorCause: Error('ProtectedData is not valid'),
+          errorCause: Error(
+            'This protected data does not contain "email:string" in its schema.'
+          ),
         })
       );
     },
     MAX_EXPECTED_WEB2_SERVICES_TIME
   );
+
   it(
     'should fail if there is no Dataset order found',
     async () => {
@@ -193,6 +199,7 @@ describe('web3mail.sendEmail()', () => {
     },
     2 * MAX_EXPECTED_BLOCKTIME + MAX_EXPECTED_WEB2_SERVICES_TIME + 5_000
   );
+
   it(
     'should successfully send email with a valid senderName',
     async () => {
@@ -209,6 +216,7 @@ describe('web3mail.sendEmail()', () => {
     },
     2 * MAX_EXPECTED_BLOCKTIME + MAX_EXPECTED_WEB2_SERVICES_TIME
   );
+
   it(
     'should successfully send email with email content size < 512 kilo-bytes',
     async () => {
@@ -229,67 +237,7 @@ describe('web3mail.sendEmail()', () => {
     },
     2 * MAX_EXPECTED_BLOCKTIME + MAX_EXPECTED_WEB2_SERVICES_TIME
   );
-  it(
-    'should fail to send email with email content size > 512 kilo-bytes',
-    async () => {
-      const desiredSizeInBytes = 520000; // 520 kilo-bytes
-      const characterToRepeat = 'A';
-      const OVERSIZED_CONTENT = characterToRepeat.repeat(desiredSizeInBytes);
 
-      const params = {
-        emailSubject: 'e2e mail object for test',
-        emailContent: OVERSIZED_CONTENT,
-        protectedData: validProtectedData.address,
-        senderName: 'Product Team',
-        workerpoolAddressOrEns: workerpoolAddress,
-      };
-      await expect(web3mail.sendEmail(params)).rejects.toThrow(
-        new WorkflowError({
-          message: 'Failed to sendEmail',
-          errorCause: Error('emailContent must be at most 512000 characters'),
-        })
-      );
-    },
-    MAX_EXPECTED_WEB2_SERVICES_TIME
-  );
-  it(
-    'should fail to send email with an invalid (too short) senderName',
-    async () => {
-      const params = {
-        emailSubject: 'e2e mail object for test',
-        emailContent: 'e2e mail content for test',
-        protectedData: validProtectedData.address,
-        senderName: 'AB',
-        workerpoolAddressOrEns: workerpoolAddress,
-      };
-      await expect(web3mail.sendEmail(params)).rejects.toThrow(
-        new WorkflowError({
-          message: 'Failed to sendEmail',
-          errorCause: Error('senderName must be at least 3 characters'),
-        })
-      );
-    },
-    MAX_EXPECTED_WEB2_SERVICES_TIME
-  );
-  it(
-    'should fail to send email with an invalid (too long) senderName',
-    async () => {
-      const params = {
-        emailSubject: 'e2e mail object for test',
-        emailContent: 'e2e mail content for test',
-        protectedData: validProtectedData.address,
-        senderName: 'A very long sender name',
-        workerpoolAddressOrEns: workerpoolAddress,
-      };
-      await expect(web3mail.sendEmail(params)).rejects.toThrow(
-        new WorkflowError({
-          message: 'Failed to sendEmail',
-          errorCause: Error('senderName must be at most 20 characters'),
-        })
-      );
-    },
-    MAX_EXPECTED_WEB2_SERVICES_TIME
-  );
   it(
     'should successfully send email with a valid label',
     async () => {
@@ -303,44 +251,6 @@ describe('web3mail.sendEmail()', () => {
       expect(sendEmailResponse.taskId).toBeDefined();
     },
     2 * MAX_EXPECTED_BLOCKTIME + MAX_EXPECTED_WEB2_SERVICES_TIME
-  );
-  it(
-    'should fail to send email with an invalid (too long) label',
-    async () => {
-      const params = {
-        emailSubject: 'e2e mail object for test',
-        emailContent: 'e2e mail content for test',
-        protectedData: validProtectedData.address,
-        label: 'ID123456789',
-        workerpoolAddressOrEns: workerpoolAddress,
-      };
-      await expect(web3mail.sendEmail(params)).rejects.toThrow(
-        new WorkflowError({
-          message: 'Failed to sendEmail',
-          errorCause: Error('label must be at most 10 characters'),
-        })
-      );
-    },
-    MAX_EXPECTED_WEB2_SERVICES_TIME
-  );
-  it(
-    'should fail to send email with an invalid (too short) label',
-    async () => {
-      const params = {
-        emailSubject: 'e2e mail object for test',
-        emailContent: 'e2e mail content for test',
-        protectedData: validProtectedData.address,
-        label: 'ID',
-        workerpoolAddressOrEns: workerpoolAddress,
-      };
-      await expect(web3mail.sendEmail(params)).rejects.toThrow(
-        new WorkflowError({
-          message: 'Failed to sendEmail',
-          errorCause: Error('label must be at least 3 characters'),
-        })
-      );
-    },
-    MAX_EXPECTED_WEB2_SERVICES_TIME
   );
 
   it(